Replace code from chapter7
- https://github.com/PacktPublishing/Hands-On-Microservices-with-Spring-Boot-and-Spring-Cloud/tree/master/Chapter07

+++ microservices/review-service/src/main/java/se/magnus/microservices/core/review/services/ReviewServiceImpl.java
@@ -1,10 +1,13 @@
 package se.magnus.microservices.core.review.services;
 
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import se.magnus.api.core.review.Review;
 import se.magnus.api.core.review.ReviewService;
 import se.magnus.microservices.core.review.persistence.ReviewEntity;
@@ -13,6 +16,9 @@ import se.magnus.util.exceptions.InvalidInputException;
 import se.magnus.util.http.ServiceUtil;
 
 import java.util.List;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
@@ -25,8 +31,12 @@ public class ReviewServiceImpl implements ReviewService {
 
     private final ServiceUtil serviceUtil;
 
+    private final Scheduler scheduler;
+
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository,
+                             ReviewMapper mapper, ServiceUtil serviceUtil) {
+        this.scheduler = scheduler;
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
@@ -34,11 +44,16 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review createReview(Review body) {
+
+        if (body.getProductId() < 1)
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
+
         try {
             ReviewEntity entity = mapper.apiToEntity(body);
             ReviewEntity newEntity = repository.save(entity);
 
-            LOG.debug("createReview: created a review entity: {}/{}", body.getProductId(), body.getReviewId());
+            LOG.debug("createReview: created a review entity: {}/{}", body.getProductId(),
+                body.getReviewId());
             return mapper.entityToApi(newEntity);
 
         } catch (DataIntegrityViolationException dive) {
@@ -47,9 +62,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public List<Review> getReviews(int productId) {
+    public Flux<Review> getReviews(int productId) {
 
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
+        LOG.info("Will get reviews for product with id={}", productId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, FINE);
+    }
+
+    protected List<Review> getByProductId(int productId) {
 
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
@@ -62,7 +84,15 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int productId) {
-        LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
+
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
+        LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}",
+            productId);
         repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
